Move in mem cache initializer to tests
